超兽不知道为什么只出现在 K 地区

+++ src/ui/OSFTrack.cc
@@ -201,7 +201,7 @@ void OSFTrackReceiver::HandleData() {
   xSum /= static_cast<float>(m_SmoothBuffer.size());
   ySum /= static_cast<float>(m_SmoothBuffer.size());
   zSum /= static_cast<float>(m_SmoothBuffer.size());
-  mouthStatus /= static_cast<float>(m_SmoothBuffer.size());
+  mouthStatus /= static_cast<int>(m_SmoothBuffer.size());
 
   emit HeadPoseUpdated(HeadPose {
     xSum, ySum, zSum,
